+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+0.11.0
+- Test command and assertions [#7](https://github.com/restbeast/restbeast/issues/7)
+
 0.10.0
 - Bytes received and sent output [#14](https://github.com/restbeast/restbeast/issues/14)
 - Bug report command `restbeast bug` [#8](https://github.com/restbeast/restbeast/issues/8)